Add files via upload
people saved highscores is now in the game. I added a few other features such as the collision with the people also. I changed how the boat moves. Now it will only go either to the left side, the right side, or the center. This is much better than before.